EOS-17083 Changes for Cleanup Command for CSM.
