+++ csm/cli/schema/csm_setup.json
@@ -269,6 +269,27 @@
         "class": "RefreshConfig",
         "is_static": false
       }
+    },
+    {
+      "name": "cleanup",
+      "description": "Perform a cleanup of all the CSM generated Files and Folders.",
+      "need_confirmation": false,
+      "permissions_tag": "update",
+      "args": [
+        {
+          "flag": "args",
+          "default": [],
+          "suppress_help": true,
+          "nargs": "?"
+        }
+      ],
+      "comm": {
+        "type": "direct",
+        "target": "csm.conf.cleanup",
+        "method": "execute",
+        "class": "Cleanup",
+        "is_static": false
+      }
     }
   ]
 }
