+++ csm/conf/cleanup.py
@@ -1,0 +1,50 @@
+# CORTX-CSM: CORTX Management web and CLI interface.
+# Copyright (c) 2020 Seagate Technology LLC and/or its Affiliates
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published
+# by the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU Affero General Public License for more details.
+# You should have received a copy of the GNU Affero General Public License
+# along with this program. If not, see <https://www.gnu.org/licenses/>.
+# For any questions about this software or licensing,
+# please email opensource@seagate.com or cortx-questions@seagate.com.
+
+from csm.conf.setup import Setup, CsmSetupError
+from csm.conf.uds import UDSConfigGenerator
+from cortx.utils.log import Log
+from cortx.utils.conf_store.conf_store import Conf
+from cortx.utils.kv_store.error import KvError
+from csm.core.blogic import const
+from csm.core.providers.providers import Response
+from csm.common.errors import CSM_OPERATION_SUCESSFUL
+
+class Cleanup(Setup):
+    """
+    Cleanup all the Files and Folders generated by csm.
+    """
+    def __init__(self):
+        super(Cleanup, self).__init__()
+        Log.info("Running Cleanup for CSM.")
+
+    async def execute(self):
+        try:
+            Conf.load(const.CSM_GLOBAL_INDEX, const.CSM_SOURCE_CONF_URL)
+        except KvError as e:
+            Log.error(f"Loading csm.conf to conf store failed {e}")
+            raise CsmSetupError("Failed to Load CSM Configuration File.")
+        self._log_cleanup()
+        self.Config.delete()
+        UDSConfigGenerator.delete()
+        return Response(output=const.CSM_SETUP_PASS, rc=CSM_OPERATION_SUCESSFUL)
+
+    def _log_cleanup(self):
+        """
+        Delete all logs
+        """
+        Log.info("Delete all logs")
+        log_path = Conf.get(const.CSM_GLOBAL_INDEX, "Log>log_path")
+        Setup._run_cmd("rm -rf " +log_path)
